fix: Potential fix for crash when JASM looks for other running instances of JASM
